Move Fq serde impl into separate file
